Restructuring bigquery and spark consumers (#73)
* Log metric values as info message when MaxBatchSizeWait is used

* Restructuring bigquery and spark consumers

(cherry picked from commit 1e158f728519f96ace56dc4060bfe34b4a614451)

+++ debezium-server-batch-sinks/src/main/java/io/debezium/server/batch/common/NoBatchSizeWait.java
@@ -20,7 +20,7 @@ import javax.inject.Named;
 @Named("NoBatchSizeWait")
 public class NoBatchSizeWait implements InterfaceBatchSizeWait {
 
-  public void waitMs(Integer numRecordsProcessed, Integer processingTimeMs) {
+  public void waitMs(long numRecordsProcessed, Integer processingTimeMs) {
   }
 
 }
